pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "indian-traffic-env"
version = "1.0.0"
description = "Indian traffic signal OpenEnv environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2",
]

[project.scripts]
server = "indian_traffic_env.server.app:main"

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[tool.hatch.build.targets.wheel]
packages = ["indian_traffic_env"]

